Sorting and Iterator used feature updated!!

+++ src/CustomerManagement/ServicePlan.java
@@ -2,11 +2,11 @@ package CustomerManagement;
 
 public enum ServicePlan {
 		SILVER(1000),GOLD(2000),DIAMOND(5000),PLATINUM(10000);
-		private int charge;
-		ServicePlan(int charge) {
+		private double charge;
+		ServicePlan(double charge) {
 			this.charge=charge;
 		}
-		public int getCharge() {
+		public double getCharge() {
 			return charge;
 		}
